Add files via upload
These screenshots show the steps for setting up Ubuntu in VirtualBox. In Power Up.png, notice that the machine name is "Linux (Initial Install) and the dropdown shows "Initial Install" -> "Current State". This is because I used the snapshot feature to take a snap of the machine with Ubuntu at its unaltered state. If a lot of changes will be made to the machine, it is best practice to have a system default available to restore.